requirements.txt.txt
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.30.0
requests>=2.31.0

# CrewAI ecosystem
crewai>=0.175.0
litellm>=1.74.0

# Google Gemini SDK
google-generativeai>=0.7.2

# Data handling
pandas>=2.0.0
openpyxl>=3.1.0
